requirements.txt
# sudo apt-get install python3-pip
# install virtual env: sudo pip3 install virtualenv

# Create :> virtualenv .my-venv
# Activate :> source .my-venv/bin/activate
# For install :> pip install -r requirements.txt

chardet==3.0.4
google-api-core==1.17.0
google-auth==1.14.2
google-cloud-container==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
idna==2.9
kubernetes==11.0.0
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
websocket-client==0.57.0